+++ cloudera-tableau/scripts/datagen/ddl/bin_flat/customer.sql
@@ -1,0 +1,8 @@
+create database if not exists ${DB};
+use ${DB};
+
+drop table if exists customer;
+
+create table customer
+stored as ${FILE}
+as select * from ${SOURCE}.customer;