+++ concourse-ci/scripts/setup_env.py
@@ -7,6 +7,7 @@ import random
 import re
 import requests
 import sys
+import base64
 from azure.storage.blob import AppendBlobService
 from azure.storage.table import TableService
 import azure.mgmt.network
@@ -31,15 +32,16 @@ def prepare_storage(settings):
 
 def render_bosh_manifest(settings):
     with open('bosh.pub', 'r') as tmpfile:
-        ssh_public_key = tmpfile.read()
+        ssh_public_key = tmpfile.read().strip()
 
     ip = netaddr.IPNetwork(settings['SUBNET_ADDRESS_RANGE_FOR_BOSH'])
     gateway_ip = str(ip[1])
     bosh_director_ip = str(ip[4])
 
     ntp_servers_maps = {
-        "AzureCloud": "0.north-america.pool.ntp.org",
-        "AzureChinaCloud": "1.cn.pool.ntp.org, 1.asia.pool.ntp.org, 0.asia.pool.ntp.org"
+        "AzureCloud": "0.pool.ntp.org, 1.pool.ntp.org",
+        "AzureChinaCloud": "1.cn.pool.ntp.org, 1.asia.pool.ntp.org, 0.asia.pool.ntp.org",
+        "AzureUSGovernment": "0.north-america.pool.ntp.org"
     }
     environment = settings["ENVIRONMENT"]
     ntp_servers = ntp_servers_maps[environment]
@@ -113,6 +115,15 @@ def render_cloud_config_manifest(settings):
         with open(cloud_config_template, 'w') as tmpfile:
             tmpfile.write(contents)
 
+def render_bosh_deployment_cmd(bosh_director_ip):
+    bosh_deployment_cmd = "deploy_bosh.sh"
+    if os.path.exists(bosh_deployment_cmd):
+        with open(bosh_deployment_cmd, 'r') as tmpfile:
+            contents = tmpfile.read()
+        contents = re.compile(re.escape("REPLACE_WITH_BOSH_DIRECTOR_IP")).sub(bosh_director_ip, contents)
+        with open(bosh_deployment_cmd, 'w') as tmpfile:
+            tmpfile.write(contents)
+
 def render_concourse_manifest(settings):
     # Render the manifest for concourse
     concourse_template = 'concourse.yml'
@@ -154,6 +165,10 @@ def get_settings():
     settings['CLIENT_ID'] = sys.argv[2]
     settings['CLIENT_SECRET'] = sys.argv[3]
 
+    print "tenant_id: {0}xxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx".format(settings['TENANT_ID'][0:4])
+    print "client_id: {0}xxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx".format(settings['CLIENT_ID'][0:4])
+    print "The length of client_secret is {0}".format(len(settings['CLIENT_SECRET']))
+
     return settings
 
 def main():
@@ -164,6 +179,7 @@ def main():
     prepare_storage(settings)
 
     bosh_director_ip = render_bosh_manifest(settings)
+    render_bosh_deployment_cmd(bosh_director_ip)
     print bosh_director_ip
 
     render_cloud_config_manifest(settings)