+++ ethereum-consortium-blockchain-network/scripts/configure-geth.sh
@@ -21,6 +21,21 @@ CONFIG_LOG_FILE_PATH="$HOMEDIR/config.log";
 cd "/home/$AZUREUSER";
 
 sudo -u $AZUREUSER /bin/bash -c "wget -N ${ARTIFACTS_URL_PREFIX}/scripts/configure-geth-azureuser.sh";
-sudo -u $AZUREUSER /bin/bash /home/$AZUREUSER/configure-geth-azureuser.sh $1 $2 $3 $4 $5 $6 $7 $8 $9 ${10} ${11} ${12} ${13} ${14} >> $CONFIG_LOG_FILE_PATH 2>&1;
-echo "===== Completed $0 =====";
-exit $?;
+
+##################################
+# Initiate loop for error checking
+##################################
+for LOOPCOUNT in `seq 1 5`; do
+	sudo -u $AZUREUSER /bin/bash /home/$AZUREUSER/configure-geth-azureuser.sh "$1" "$2" "$3" "$4" "$5" "$6" "$7" "$8" "$9" "${10}" "${11}" "${12}" "${13}" "${14}" "${15}">> $CONFIG_LOG_FILE_PATH 2>&1;
+	if [ $? -ne 0 ]; then
+		echo "Command failed on try $LOOPCOUNT, retrying..." >> $CONFIG_LOG_FILE_PATH;
+		sleep 5;
+		continue;
+	else
+		echo "======== Deployment successful! ======== " >> $CONFIG_LOG_FILE_PATH;
+		exit 0;
+	fi
+done
+
+echo "One or more commands failed after 5 tries. Deployment failed." >> $CONFIG_LOG_FILE_PATH;
+exit 1