+++ moodle-cluster-ubuntu/install_frontend.sh
@@ -41,12 +41,12 @@ curl -k --max-redirs 10 https://github.com/moodle/moodle/archive/$moodleVersion.
 unzip moodle.zip
 mv moodle-$moodleVersion moodle
 
-# Install Office 365 plugins if asked for
+# install Office 365 plugins if asked for
 if [ "$installOfficePlugins" = "True" ]; then
     curl -k --max-redirs 10 https://github.com/Microsoft/o365-moodle/archive/$moodleVersion.zip -L -o o365.zip
     unzip o365.zip
     
-# The plugins below are not required for new installations
+    # The plugins below are not required for new installations
     rm -rf o365-moodle-$moodleVersion/blocks/onenote
     rm -rf o365-moodle-$moodleVersion/local/m*
     rm -rf o365-moodle-$moodleVersion/local/o365docs
@@ -56,11 +56,11 @@ if [ "$installOfficePlugins" = "True" ]; then
     rm -rf o365-moodle-$moodleVersion/user/profile/
     rm -rf o365-moodle-$moodleVersion/repository/onenote	
     
-# Copy office plugins folder into moodle  and remove unzipped folder
+    #Copy office plugins to moodle and remove office unzipped folder
+   
     cp -r o365-moodle-$moodleVersion/* moodle
     rm -rf o365-moodle-$moodleVersion
 fi
-
 
 # make the moodle directory writable for owner
 chown -R www-data moodle