+++ moodle-singlevm-ubuntu/install_lamp_moodle.sh
@@ -42,7 +42,7 @@ if [ "$installOfficePlugins" = "True" ]; then
     curl -k --max-redirs 10 https://github.com/Microsoft/o365-moodle/archive/$moodleVersion.zip -L -o o365.zip
     unzip o365.zip
     
-# The plugins below are not required for new installations
+    # The plugins below are not required for new installations
     rm -rf o365-moodle-$moodleVersion/blocks/onenote
     rm -rf o365-moodle-$moodleVersion/local/m*
     rm -rf o365-moodle-$moodleVersion/local/o365docs
@@ -51,8 +51,8 @@ if [ "$installOfficePlugins" = "True" ]; then
     rm -rf o365-moodle-$moodleVersion/mod/assign
     rm -rf o365-moodle-$moodleVersion/user/profile/
     rm -rf o365-moodle-$moodleVersion/repository/onenote	
+    cp -r o365-moodle-$moodleVersion/* moodle
     
-# Copy office plugins folder into moodle  and remove unzipped folder
     cp -r o365-moodle-$moodleVersion/* moodle
     rm -rf o365-moodle-$moodleVersion
 fi